Add ParaView 5.9.0 sources

+++ VTK/Rendering/OpenVR/vtkOpenVRModel.cxx
@@ -52,7 +52,7 @@ vtkOpenVRModel::~vtkOpenVRModel()
   this->ModelVBO = 0;
 }
 
-//----------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 void vtkOpenVRModel::PrintSelf(ostream& os, vtkIndent indent)
 {
   this->Superclass::PrintSelf(os, indent);
@@ -244,6 +244,6 @@ void vtkOpenVRModel::SetRayLength(double length)
 
 void vtkOpenVRModel::SetRayColor(double r, double g, double b)
 {
-  float color[] = { r, g, b };
+  float color[] = { static_cast<float>(r), static_cast<float>(g), static_cast<float>(b) };
   this->Ray->SetColor(color);
 }
